Deliverable 1 update by Karan
UPDATE://
It is currently 1:44 AM – Thursday morning
I added the missing part to the front matter (check the bottom of that
page)
I did the executive overview, it is a page and it pretty much covers it
all. The only issue is that it feels bloated, so we need to chat about
what is really needed there.
I reviewed the whole document and cleaned it up as much as possible. Let
me know if any changes are needed.
After rereading the PDF, it said responsibilities “should” be assigned
for a different person every deliverable. I need clarification on that.
I am talking about the highlighted the text in this document.
- Karan